expose BrowserCore objects from IWebBrowser
The immediate purpose here is to make it easier to add local JS object
binding. Rather than have each WebView implement the interface on its own,
we can simply have the BrowserCore implement it, and expose that. That
also allows us to keep parts of the interface "internal", which wouldn't
be possible if it was all exposed through IWebBrowser. If we like this
change, we might get rid of some of the IWebBrowser methods that are just
wrappers, and instead rely on shared code in BrowserCore.