Fixed what Keys managedCefBrowserAdapter send
Problem: Not all input-keys was sent to handled in Javascript, especially keycombinations like Strg + X, Strg + V and so on wasn't handled. Now, with the fix, this is no longer a problem. Also the keys, special handled in the original version, are no longer needed and so i removed them.