Startpage "about:blank" removed
Problem: The Back-Button doesn't work on the right way, because the with about-blank he's activated (like you can press the button) but doesn't load anything, because about:blank isn't a real site. So the running-statusbar always running, but no progress is visible. Also there was the possibility, that the real Url never loaded. Don't know how these works together, but with an empty startpage, this problem is gone.