Add late binding to Javascript object and array
Added late binding to JavaScript object, so that objects with cyclic
reference don't cause stack overflow (updated example to prove that).

For consistency, modified method calls to return a true object, just
like property calls.

Add IsNull property to JavascriptObject  to indicate if the object is
null, so that on browser side we don't need to create a CEF object.