Added support for drag & drop from input fields
There was a problem with the previous implementation. It worked correctly for many use cases:

- Dragging from a web app to a WPF control.
- Dragging from a WPF control to a web app.
- Dragging _html content_ from a web app to an input field

**But**: when dragging text from one input field to another, the text would vanish (!) when you drop it.

We debugged this here quite extensively. It turned out that if you have _set_ the `FragmentHtml` (even though you've set it to an empty string), it behaves like this. The solution is to avoid setting it if it is empty.