Nupkg Restructure
Resolves #2156

Nuget packages updated to use ItemGroup None instead of Copy target
Nuget props restructure to support Win32 target
x86 and Win32 targets are exactly the same in content, they differ in name only, so switched to using an <Otherwise/> block
Remove AnyCpu <When/> option from props files and consolidate with x86 and Win32. Set Private to false and rely on the
<None/> entires to copy the files to the output folder

Ideally would move the OffScreen, WinForms and WPF targets file content to the props (as they're actually props now), unfortunately
we need guaranteed ordering and Nuget 2.8.x doesn't provide that, if we leave them as .targets it'll happen late enough to make sure
the cef.redist list of files is provided.
https://github.com/NuGet/Home/issues/4229#issuecomment-271387190