Fix non-thread-safety of TrySetResultAsync
The extension method TrySetResultAsync has non-deterministic results when called twice on the same instance, even on the same thread: it creates one task for each call, which may not run in their creation order.
This commit adds a wrapper around TaskCompletionSource, that ensures that only the first call to TrySetResultAsync is really executed.
It also updates the test case that could be violated with the previous behavior (but that was hard to trigger with only one test).