Bug fix: handle type=crashpad-handler
Cherry-pick of #2449 which was approved by Alex.

I was seeing instances of the subprocess being spawned this way when running the CefSharp.Wpf.Example project. Debugging showed me that no such argument was being passed to the process. This PR works around the issue by opting out from getting the parent process ID for such processes.