Cef.h - Remove ThrowIfCefNotInitialized
As IsInitialized is set when CefInitialize returns and it appears to be possible the OnContextInitialized is called
before it returns then this exception wasn't going to work.

Testing with Cef.GetGlobalRequestContext and Cef.GetGlobalCookieManager and they both return null if called before Cef.Initialize
So it shouldn't be possible to actually access the cookie manager/RequestContext. Improved the xml doc

Resolves #2600