Nuget - Additional PackageReference set of nupkg files
**WORK IN PROGRESS**

Unfortunately the new chromiumembeddedframework.redist.win-x86/chromiumembeddedframework.redist.win-x64 packages
aren't usable yet, the files are copied to the relevant natives folder .net core just doesn't load the dlls.
I suspect if we were to p/invoke the dll then it would be loaded correctly, might need to raise this with .Net Core team.
So for now just using the older cef.redist packages, which expose props that can be used in any other package.
the new CefSharp.Common.win packages will copy the files using the .targets file as before, difference being PlatformTarget
is used instead of Platform so we can detect the actual platform of the project not the solution.

AnyCPU is not currently supported, x86/x64 only, Win32 might work, not tested though.
Not clear that AnyCpu will be possible, more digging to do on that front.

Issue #2795