ClickOnce - Load CefSharp.dll from parent directory if not in current folder
When publishing with ClickOnce targeting AnyCPU the x64 folder is missing a copy of CefSharp.dll which is required by the BrowserSubprocess.
The file is copied correctly when the projects builds normally, so there's some quirk of the Publish process.

Temp workaround

Issue as described at https://github.com/cefsharp/CefSharp/issues/3319#issuecomment-791783346