Update docs for IResourceRequestHandlerFactory (#3998)
I have just discovered that when HasHandlers returns false, the GetResourceRequestHandler method is never executed.  Presumably this is an optimisation which happens inside of Chromium, as I can see that the HasHandlers property value goes all way back to the C++ native project.

I was about to suggest removing this optimisation because it'd be trivial to optimise and return early inside the default factory impl in C# world.  If it goes all the way down to Chromium then that's less easy and so the next best thing would be to document it.