CefSharp.Dom - Don't Dispose DevToolsContext
Disposing of the DevToolsContext is causing problems currently as sending the DevTools protocol messages
to the browser is happening after the CefBrowserHost instance has been Disposed.

A separate method of disposing of the DevToolsContext will need to be added that does a cleanup
without sending the commands.

Issue https://github.com/cefsharp/CefSharp.Dom/issues/41