Add Compatibility with Central Package Management
When a runtime identifier is NOT chosen, the default Nuget logic is to reference all possible compilations for compatibility purposes. However if Central Package Management, introduced in Nuget 6.2 ( https://devblogs.microsoft.com/nuget/introducing-central-package-management/ ) is enabled referencing the package directly in a targets file will causes an error. Additional since it is only imported at compile time, it triggers a infinite restore loop in visual studio.